pyproject.toml
[tool.poetry]
name = "valenspy"
version = "0.1.0"
description = "A Python package to validate gridded model data."
authors = ["Kobe Vandelanotte <kobe.vandelanotte@meteo.be>",
	   " ... "]
license = ","
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
xarray = "^2024.2.0"
dask = "^2024.3.1"
netCDF4 = "^1.6.5"
pandas = "^2.2.1"
shapely = "^2.0.3"
xarray-datatree = "^0.0.14"
matplotlib = "^3.8.3"
scipy = "^1.13.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.2"
Sphinx = "^7.2.6"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
